requirements.txt
requests
dnspython
ipwhois
Flask
